requirements.txt
streamlit==1.37.1
trafilatura==1.8.1
readability-lxml==0.8.1
spacy==3.7.4
scikit-learn==1.4.2
sentence-transformers==2.7.0
networkx==3.2.1
numpy==1.26.4
pandas==2.2.2
beautifulsoup4==4.12.3
tqdm==4.66.4
python-dotenv==1.0.1
requests==2.32.3
lxml==5.2.2
urllib3==2.2.2
pyyaml==6.0.2
